refactor: `User` API `DTO` 위치 변경
+ refactor: 예외들 `private constructor` 로 변경
