+++ src/main/java/com/palettee/user/controller/dto/response/users/GetUserArchiveResponse.java
@@ -1,4 +1,4 @@
-package com.palettee.user.controller.dto.response;
+package com.palettee.user.controller.dto.response.users;
 
 import com.palettee.archive.domain.*;
 import java.util.*;
