+++ src/main/java/com/palettee/user/controller/dto/response/users/GetUserGatheringResponse.java
@@ -1,4 +1,4 @@
-package com.palettee.user.controller.dto.response;
+package com.palettee.user.controller.dto.response.users;
 
 import com.palettee.gathering.domain.*;
 import java.util.*;
