+++ src/main/java/com/palettee/user/exception/InvalidJobGroupException.java
@@ -6,7 +6,7 @@ public class InvalidJobGroupException extends PaletteException {
 
     public static PaletteException EXCEPTION = new InvalidJobGroupException();
 
-    public InvalidJobGroupException() {
+    private InvalidJobGroupException() {
         super(ErrorCode.INVALID_JOB_GROUP);
     }
 }
