+++ src/main/java/com/palettee/user/exception/NotOwnUserException.java
@@ -6,7 +6,7 @@ public class NotOwnUserException extends PaletteException {
 
     public static PaletteException EXCEPTION = new NotOwnUserException();
 
-    public NotOwnUserException() {
+    private NotOwnUserException() {
         super(ErrorCode.NOT_OWN_USER);
     }
 }
