+++ src/main/java/com/palettee/user/exception/UserAccessException.java
@@ -1,13 +1,12 @@
 package com.palettee.user.exception;
 
-import com.palettee.global.exception.ErrorCode;
-import com.palettee.global.exception.PaletteException;
+import com.palettee.global.exception.*;
 
 public class UserAccessException extends PaletteException {
 
     public static final PaletteException EXCEPTION = new UserAccessException();
 
-    public UserAccessException() {
+    private UserAccessException() {
         super(ErrorCode.NOT_ACCESS);
     }
