+++ src/main/java/com/palettee/user/exception/UserNotFoundException.java
@@ -1,13 +1,12 @@
 package com.palettee.user.exception;
 
-import com.palettee.global.exception.ErrorCode;
-import com.palettee.global.exception.PaletteException;
+import com.palettee.global.exception.*;
 
 public class UserNotFoundException extends PaletteException {
 
     public static final PaletteException EXCEPTION = new UserNotFoundException();
 
-    public UserNotFoundException() {
+    private UserNotFoundException() {
         super(ErrorCode.USER_NOT_FOUND);
     }
 }
