+++ src/test/java/com/palettee/user/controller/BasicRegisterControllerTest.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.databind.*;
 import com.palettee.global.exception.*;
 import com.palettee.global.security.jwt.exceptions.*;
 import com.palettee.global.security.jwt.utils.*;
-import com.palettee.user.controller.dto.request.*;
+import com.palettee.user.controller.dto.request.users.*;
 import com.palettee.user.domain.*;
 import com.palettee.user.repository.*;
 import jakarta.transaction.*;
