+++ src/test/java/com/palettee/user/service/BasicRegisterServiceTest.java
@@ -3,8 +3,8 @@ package com.palettee.user.service;
 import static org.assertj.core.api.Assertions.*;
 
 import com.palettee.portfolio.repository.*;
-import com.palettee.user.controller.dto.request.*;
-import com.palettee.user.controller.dto.response.*;
+import com.palettee.user.controller.dto.request.users.*;
+import com.palettee.user.controller.dto.response.users.*;
 import com.palettee.user.domain.*;
 import com.palettee.user.exception.*;
 import com.palettee.user.repository.*;
